ATS Final/requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.2
python-dotenv>=1.0.0
python-docx>=1.0.1
pypdf2>=3.0.1
requests>=2.31.0
python-magic>=0.4.27
typing-extensions>=4.8.0
pydantic>=2.5.2
argparse>=1.4.0